Update CI and replace deprecated np.float and np.bool (#127)
* replace deprecated np.float and np.bool

* amend readme

* Python 3.10 tests instead of 3.7

* enable py3.10 test

* remove deprecated lgtm badge

+++ lettuce/stencils.py
@@ -9,7 +9,7 @@ __all__ = ["Stencil", "D1Q3", "D2Q9", "D3Q15", "D3Q19", "D3Q27"]
 class Stencil:
     e: np.ndarray = np.array([])
     w: np.ndarray = np.array([])
-    cs: np.float = 0.0
+    cs: float = 0.0
     opposite: [int] = []
 
     @classmethod
